refactor: move class to new package
